+++ python/cudf/dataframe/dataframe.py
@@ -1379,16 +1379,16 @@ class DataFrame(object):
                 rcats = rhs[name].cat.categories
                 if how == 'rhs':
                     cats = rcats
-                    lhs[name] = (lhs[name].cat.set_categories(cats)
+                    lhs[name] = (lhs[name].cat._set_categories(cats)
                                  .fillna(-1))
                 elif how in ['inner', 'outer']:
                     # Do the join using the union of categories from both side.
                     # Adjust for inner joins afterwards
                     cats = sorted(set(lcats) | set(rcats))
-                    lhs[name] = (lhs[name].cat.set_categories(cats)
+                    lhs[name] = (lhs[name].cat._set_categories(cats)
                                  .fillna(-1))
                     lhs[name] = lhs[name]._column.as_numerical
-                    rhs[name] = (rhs[name].cat.set_categories(cats)
+                    rhs[name] = (rhs[name].cat._set_categories(cats)
                                  .fillna(-1))
                     rhs[name] = rhs[name]._column.as_numerical
                 col_cats[name] = cats
@@ -1398,16 +1398,16 @@ class DataFrame(object):
                 rcats = rhs[name].cat.categories
                 if how == 'left':
                     cats = lcats
-                    rhs[name] = (rhs[name].cat.set_categories(cats)
+                    rhs[name] = (rhs[name].cat._set_categories(cats)
                                  .fillna(-1))
                 elif how in ['inner', 'outer']:
                     # Do the join using the union of categories from both side.
                     # Adjust for inner joins afterwards
                     cats = sorted(set(lcats) | set(rcats))
-                    lhs[name] = (lhs[name].cat.set_categories(cats)
+                    lhs[name] = (lhs[name].cat._set_categories(cats)
                                  .fillna(-1))
                     lhs[name] = lhs[name]._column.as_numerical
-                    rhs[name] = (rhs[name].cat.set_categories(cats)
+                    rhs[name] = (rhs[name].cat._set_categories(cats)
                                  .fillna(-1))
                     rhs[name] = rhs[name]._column.as_numerical
                 col_cats[name] = cats
@@ -1598,23 +1598,23 @@ class DataFrame(object):
             if how == 'left':
                 cats = lcats
                 rhs[idx_col_name] = (rhs[idx_col_name].cat
-                                                      .set_categories(cats)
+                                                      ._set_categories(cats)
                                                       .fillna(-1))
             elif how == 'right':
                 cats = rcats
                 lhs[idx_col_name] = (lhs[idx_col_name].cat
-                                                      .set_categories(cats)
+                                                      ._set_categories(cats)
                                                       .fillna(-1))
             elif how in ['inner', 'outer']:
                 cats = sorted(set(lcats) | set(rcats))
 
                 lhs[idx_col_name] = (lhs[idx_col_name].cat
-                                                      .set_categories(cats)
+                                                      ._set_categories(cats)
                                                       .fillna(-1))
                 lhs[idx_col_name] = lhs[idx_col_name]._column.as_numerical
 
                 rhs[idx_col_name] = (rhs[idx_col_name].cat
-                                                      .set_categories(cats)
+                                                      ._set_categories(cats)
                                                       .fillna(-1))
                 rhs[idx_col_name] = rhs[idx_col_name]._column.as_numerical
 
@@ -2140,7 +2140,7 @@ class DataFrame(object):
 
         Parameters
         ----------
-        data : numpy structured dtype or recarray
+        data : numpy structured dtype or recarray of ndim=2
         index : str
             The name of the index column in *data*.
             If None, the default index is used.
@@ -2151,12 +2151,31 @@ class DataFrame(object):
         -------
         DataFrame
         """
-        names = data.dtype.names if columns is None else columns
+        if data.ndim != 1 and data.ndim != 2:
+            raise ValueError("records dimension expected 1 or 2 but found {!r}"
+                             .format(data.ndim))
+
+        num_cols = len(data[0])
+        if columns is None and data.dtype.names is None:
+            names = [i for i in range(num_cols)]
+
+        elif data.dtype.names is not None:
+            names = data.dtype.names
+
+        else:
+            if len(columns) != num_cols:
+                msg = "columns length expected {!r} but found {!r}"
+                raise ValueError(msg.format(num_cols, len(columns)))
+            names = columns
+
         df = DataFrame()
-        for k in names:
-            # FIXME: unnecessary copy
-            df[k] = Series(np.ascontiguousarray(data[k]),
-                           nan_as_null=nan_as_null)
+        if data.ndim == 2:
+            for i, k in enumerate(names):
+                df[k] = Series(data[:, i], nan_as_null=nan_as_null)
+        elif data.ndim == 1:
+            for k in names:
+                df[k] = Series(data[k], nan_as_null=nan_as_null)
+
         if index is not None:
             indices = data[index]
             return df.set_index(indices.astype(np.int64))
@@ -2189,12 +2208,12 @@ class DataFrame(object):
         else:
             if len(columns) != data.shape[1]:
                 msg = "columns length expected {!r} but found {!r}"
-                raise ValueError(msg.format(data.ndim, len(columns)))
+                raise ValueError(msg.format(data.shape[1], len(columns)))
             names = columns
 
         if index is not None and len(index) != data.shape[0]:
             msg = "index length expected {!r} but found {!r}"
-            raise ValueError(msg.format(data.ndim, len(columns)))
+            raise ValueError(msg.format(data.shape[0], len(index)))
 
         df = DataFrame()
         data = data.transpose()  # to mimic the pandas behaviour